Apis Serviços Veiculos

+++ src/main/java/com/tinnova/coder/service/VeiculoService.java
@@ -1,11 +1,13 @@
 package com.tinnova.coder.service;
 
+import com.tinnova.coder.controller.dto.VeiculoDto;
 import com.tinnova.coder.domain.Veiculo;
 import com.tinnova.coder.repository.VeiculoRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class VeiculoService {
@@ -24,25 +26,86 @@ public class VeiculoService {
         return veiculoRepository.findById(id);
     }
 
-    public Veiculo save(Veiculo veiculo) {
-        return veiculoRepository.save(veiculo);
+    public Optional<VeiculoDto> updateVeiculo(Long id, VeiculoDto veiculoDto) {
+        Optional<Veiculo> veiculoExistente = veiculoRepository.findById(id);
+
+        if (veiculoExistente.isPresent()) {
+            // Recupera o veículo existente
+            Veiculo veiculoAtual = veiculoExistente.get();
+
+            // Atualiza os campos do veículo com os dados do DTO
+            if (veiculoDto.getMarca() != null) veiculoAtual.setMarca(veiculoDto.getMarca());
+            if (veiculoDto.getAno() != null) veiculoAtual.setAno(veiculoDto.getAno());
+            if (veiculoDto.getCor() != null) veiculoAtual.setCor(veiculoDto.getCor());
+
+            // Define o ID com o valor passado na URL
+            veiculoAtual.setId(id);
+
+            // Salva o veículo atualizado
+            Veiculo veiculoAtualizado = veiculoRepository.save(veiculoAtual);
+
+            // Retorna o veículo atualizado como DTO
+            return Optional.of(new VeiculoDto(veiculoAtualizado));
+        }
+        return Optional.empty();  // Retorna Optional vazio caso o veículo não seja encontrado
     }
 
-    public Veiculo partialUpdate(Long id, Veiculo veiculo) {
-        Veiculo veiculoExistente = veiculoRepository.findById(id).orElseThrow();
+
+    public VeiculoDto save(Veiculo veiculo) {
+        return new VeiculoDto(veiculoRepository.save(veiculo));
+    }
+
+    public VeiculoDto partialUpdate(Long id, VeiculoDto veiculo) {
+        Veiculo veiculoExistente = veiculoRepository.findById(id).orElseThrow(() -> new IllegalStateException("Veículo não encontrado"));
         // Atualiza parcialmente os campos
         if (veiculo.getMarca() != null) veiculoExistente.setMarca(veiculo.getMarca());
         if (veiculo.getAno() != null) veiculoExistente.setAno(veiculo.getAno());
         if (veiculo.getCor() != null) veiculoExistente.setCor(veiculo.getCor());
-        return veiculoRepository.save(veiculoExistente);
+        return new VeiculoDto(veiculoRepository.save(veiculoExistente));
     }
 
-    public void delete(Long id) {
-        veiculoRepository.deleteById(id);
+    public Optional<VeiculoDto> delete(Long id) {
+        Optional<Veiculo> veiculoExistente = veiculoRepository.findById(id);
+
+        // Se o veículo existir, exclui ele
+        if (veiculoExistente.isPresent()) {
+            veiculoRepository.deleteById(id);
+            return Optional.of(new VeiculoDto(veiculoExistente.get()));  // Retorna o Veículo como DTO
+        }
+
+        // Caso o veículo não exista, retorna Optional vazio
+        return Optional.empty();
     }
 
-    public List<Veiculo> findByMarcaAnoCor(String marca, Integer ano, String cor) {
-        return veiculoRepository.findByMarcaOrAnoOrCor(marca, ano, cor);
+    public Optional<VeiculoDto> getVeiculoById(Long id) {
+        Optional<Veiculo> veiculoExistente = veiculoRepository.findById(id);
+
+        // Se o veículo existir, exclui ele
+        if (veiculoExistente.isPresent()) {
+            return Optional.of(new VeiculoDto(veiculoExistente.get()));
+        }
+
+        return Optional.empty();
+    }
+
+
+    public List<VeiculoDto> findVeiculos(String marca, Integer ano, String cor) {
+        // Verifica se qualquer um dos parâmetros é diferente de null
+        if (marca != null || ano != null || cor != null) {
+            return findByMarcaAnoCor(marca, ano, cor);
+        }
+        // Se todos os parâmetros forem null, retorna todos os veículos
+        return findAll()
+                .stream()
+                .map(VeiculoDto::new)
+                .collect(Collectors.toList());
+    }
+
+    public List<VeiculoDto> findByMarcaAnoCor(String marca, Integer ano, String cor) {
+        return veiculoRepository.findByMarcaOrAnoOrCor(marca, ano, cor)
+                .stream()
+                .map(VeiculoDto::new)  // Converte cada Veiculo para VeiculoDto
+                .collect(Collectors.toList());
     }
 }
